pyproject.toml
[tool.ruff]
# Exclude a variety of commonly ignored directories.
extend-exclude = [
    ".direnv",
    ".git",
    ".ruff_cache",
    "node_modules",
    "migrations",
]
line-length = 120
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
fixable = [
    "ALL",
]
unfixable = []
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.setuptools]
py-modules = []

[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "Django>=5.2",
    "celery",
    "celery-progress",
    "celery[redis]",
    "channels",
    "channels_redis",
    "dj-stripe>=2.9",
    "django-allauth[socialaccount,mfa]",
    "django-anymail[mailersend]", # email services
    "django-celery-beat",
    "django-environ",
    "django-health-check",
    "django-hijack",
    "django-htmx",
    "django-template-partials",
    "django-vite>=3.1.0",
    "django-waffle",
    "djangorestframework",
    "djangorestframework-api-key",
    "docutils",
    "drf-spectacular",
    "httpx>=0.28.1",
    "litellm>=1.65.1",
    "markdown",
    "nh3>=0.2.21",
    "openai",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "psycopg2-binary",
    "pydantic-ai>=1.0",
    "python-dateutil",
    "redis",
    "requests",
    "sentry-sdk[django]>=2.33.2",
    "stripe<12.0",
    "whitenoise[brotli]",
    "xlsxwriter",
]

[dependency-groups]
dev = [
    "channels[daphne]",
    "pre-commit",
    "ruff>=0.12.8",
    "pegasus-cli>=0.8",
]
prod = [
    "gevent",
    "gunicorn",
    "uvicorn[standard]",
]
